requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
openai>=1.0.0
numpy
python-dotenv
